requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
segmentation_models_pytorch==0.5.0
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
